websearch/mcp/gigachat-summary/pyproject.toml
[project]
name = "gigachat-summary-mcp"
version = "0.1.0"
description = "MCP сервер summarize(text, style, max_chars) поверх GigaChat"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=0.4.2",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
